Fix race condition in drawing
The root cause is that there are two systems at work:
1. A keypress callback happens on one thread
2. The UI has a sync timer happening on another

When a keypress happens, the state of the stream deck is updated.
When the sync timer fires, it will set brightness and page number.
If the page number is different, this will trigger an even that switches
pages, which in turn will render the images again.

In addition to rendering images when switching pages, the button text
field will be updated, again triggering an event that renders the
buttons.  This means there are potentially three things trying to update
the buttons at the same time.

At the heart of all of this is the fact that the UI event handlers does
not know if the UI or the streamdeck is causing the state change.

The change here is quite simple - verify if the API state is different
before saving it. This is a sensible API change regardless of this
particular problem. It avoids uncessary writes to the JSON file too,
but most importantly, avoids re-entering the rendering function
(which was redundant anyways as the buttons have already been updated).

Technically, this is still not thread safe, as the streamdeck callback thread
and the UI thread could in theory run at exactly the same time.

In practice, this would be quite unlikely to happen; and this issue
has always existed in the code.

In addition, synchronization code has been added to only allow access to
one streamdeck at a time while updating the images.

This should fix the scrambled image issue.

Further to this - it may make more sense to marshal the button event
using a signal over the Qt threading model. This should allow events
and other things to happen on the same thread.