Remove python3-xlib dependency
It appears python3-xlib is an unmaintained historic attempt to support python3.
python-xlib has been maintained and supports python3+
Some testing still needs to be done that this works on a clean install.
When removing python3-xlib, with python-xlib already installed breaks the app.
Reinstalling python-xlib fixes it. It looks like the uninstall must be removing
some common files.